refactor: restructure components into logical subdirectories
