+++ src/app/audio/components/browser/index.ts
@@ -1,0 +1,2 @@
+export { SoundBrowserComponent } from './sound-browser.component';
+export { PreviewClipComponent } from './preview-clip.component';
