+++ src/app/audio/components/browser/sound-browser.component.ts
@@ -7,8 +7,8 @@ import { MatInputModule } from '@angular/material/input';
 import { MatChipsModule } from '@angular/material/chips';
 import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
 import { MatTooltipModule } from '@angular/material/tooltip';
-import { SoundLibraryService } from '../services/sound-library.service';
-import { SoundLibraryItem, SoundCategory } from '../utils/sound-library';
+import { SoundLibraryService } from '../../services/sound-library.service';
+import { SoundLibraryItem, SoundCategory } from '../../utils/sound-library';
 
 @Component({
     selector: 'sound-browser',
