+++ src/app/audio/components/dialogs/index.ts
@@ -1,0 +1,4 @@
+export { ConfirmDialogComponent } from './confirm-dialog.component';
+export { LoadArrangementDialogComponent } from './load-arrangement-dialog.component';
+export { RecordingDialogComponent } from './recording-dialog.component';
+export { SaveArrangementDialogComponent } from './save-arrangement-dialog.component';
