+++ src/app/audio/components/editor/audio-editor.component.css
