+++ src/app/audio/components/editor/bottom-panel.component.css
