+++ src/app/audio/components/editor/index.ts
@@ -1,0 +1,3 @@
+export { AudioEditorComponent } from './audio-editor.component';
+export { BottomPanelComponent } from './bottom-panel.component';
+export { LoopRegionComponent } from './loop-region.component';
