+++ src/app/audio/components/index.ts
@@ -1,0 +1,13 @@
+// Main barrel export for all audio components
+
+// Editor components
+export * from './editor';
+
+// Timeline components  
+export * from './timeline';
+
+// Browser components
+export * from './browser';
+
+// Dialog components
+export * from './dialogs';
