+++ src/app/audio/components/timeline/clip.component.css
