+++ src/app/audio/components/timeline/track-header.component.css
