+++ src/app/audio/components/timeline/track-lane.component.css
