+++ src/app/audio/components/timeline/track.component.css
